Split module into multiple files
