+++ truckersmp_cli/__init__.py
@@ -4,28 +4,18 @@ Module for truckersmp-cli main script.
 Licensed under MIT.
 """
 
+from truckersmp_cli import utils
+from truckersmp_cli import downloads
+from truckersmp_cli import variables
+
 import argparse
-import ctypes
-import hashlib
-import html.parser
-import http.client
-import io
 import json
 import locale
 import logging
 import os
-import platform
-import shutil
 import signal
 import subprocess as subproc
 import sys
-import tarfile
-import time
-import urllib.parse
-import urllib.request
-from getpass import getuser
-from gettext import ngettext
-from zipfile import ZipFile
 
 pkg_resources_is_available = False
 try:
@@ -34,466 +24,11 @@ try:
 except ImportError:
     pass
 
-vdf_is_available = False
-try:
-    import vdf
-    vdf_is_available = True
-except ImportError:
-    pass
-
-
-class URL:
-    """URLs."""
-
-    project = "https://github.com/lhark/truckersmp-cli"
-    dlurl = "download.ets2mp.com"
-    dlurlalt = "failover.truckersmp.com"
-    listurl = "https://update.ets2mp.com/files.json"
-    steamcmdlnx = "https://steamcdn-a.akamaihd.net/client/installer/steamcmd_linux.tar.gz"
-    steamcmdwin = "https://steamcdn-a.akamaihd.net/client/installer/steamcmd.zip"
-    raw_github = "raw.githubusercontent.com"
-    d3dcompilerpath = "/ImagingSIMS/ImagingSIMS/master/Redist/x64/d3dcompiler_47.dll"
-    truckersmp_api = "https://api.truckersmp.com/v2/version"
-    truckersmp_stats = "https://stats.truckersmp.com"
-    truckersmp_downgrade_help = "https://truckersmp.com/kb/26"
-    issueurl = project + "/issues"
-    release = project + "/raw/master/RELEASE"
-    rel_tarxz_tmpl = project + "/releases/download/{0}/truckersmp-cli-{0}.tar.xz"
-
-
-class Dir:
-    """Directories."""
-
-    XDG_DATA_HOME = os.getenv("XDG_DATA_HOME", os.path.expanduser("~/.local/share"))
-    default_gamedir = {
-        "ats": os.path.join(
-          XDG_DATA_HOME, "truckersmp-cli/American Truck Simulator/data"),
-        "ets2": os.path.join(
-          XDG_DATA_HOME, "truckersmp-cli/Euro Truck Simulator 2/data"),
-    }
-    default_prefixdir = {
-        "ats": os.path.join(
-          XDG_DATA_HOME, "truckersmp-cli/American Truck Simulator/prefix"),
-        "ets2": os.path.join(
-          XDG_DATA_HOME, "truckersmp-cli/Euro Truck Simulator 2/prefix"),
-    }
-    default_moddir = os.path.join(XDG_DATA_HOME, "truckersmp-cli/TruckersMP")
-    default_protondir = os.path.join(XDG_DATA_HOME, "truckersmp-cli/Proton")
-    steamcmddir = os.path.join(XDG_DATA_HOME, "truckersmp-cli/steamcmd")
-    steamcmdpfx = os.path.join(steamcmddir, "pfx")
-    dllsdir = os.path.join(XDG_DATA_HOME, "truckersmp-cli/dlls")
-    scriptdir = os.path.dirname(os.path.realpath(__file__))
-
-
-class File:
-    """Files."""
-
-    loginvdf_inner = "config/loginusers.vdf"
-    # known paths for [steam installation directory]/config/loginusers.vdf
-    loginusers_paths = [
-        # Official (Valve) version
-        os.path.join(Dir.XDG_DATA_HOME, "Steam", loginvdf_inner),
-        # Debian-based systems, old path
-        os.path.join(os.path.expanduser("~/.steam"), loginvdf_inner),
-        # Debian-based systems, new path
-        os.path.join(os.path.expanduser("~/.steam/debian-installation"), loginvdf_inner),
-    ]
-    proton_json = os.path.join(Dir.scriptdir, "proton.json")
-    inject_exe = os.path.join(Dir.scriptdir, "truckersmp-cli.exe")
-    overlayrenderer_inner = "ubuntu12_64/gameoverlayrenderer.so"
-    d3dcompiler_47 = os.path.join(Dir.dllsdir, "d3dcompiler_47.dll")
-    d3dcompiler_47_md5 = "b2cc65e1930e75f563078c6a20221b37"
-    sdl2_soname = "libSDL2-2.0.so.0"
-
-
-class AppId:
-    """Steam AppIDs."""
-
-    game = {
-        "ats":          270880,        # https://steamdb.info/app/270880/
-        "ets2":         227300,        # https://steamdb.info/app/227300/
-    }
-    proton = {}
-
-
-class DowngradeHTMLParser(html.parser.HTMLParser):
-    """Extract downgrade information from HTML code at stats.truckersmp.com."""
-
-    _data = {"ets2": False, "ats": False}
-    _is_downgrade_node = False
-
-    def handle_starttag(self, tag, attrs):
-        """HTML start tag handler."""
-        for attr in attrs:
-            if (attr[0] == "href"
-                    and len(attr) > 1
-                    and attr[1] == URL.truckersmp_downgrade_help):
-                self._is_downgrade_node = True
-                break
-
-    def handle_endtag(self, tag):
-        """HTML end tag handler."""
-        self._is_downgrade_node = False
-
-    def handle_data(self, data):
-        """HTML data handler."""
-        if self._is_downgrade_node:
-            if "ETS2" in data:
-                self._data["ets2"] = True
-            if "ATS" in data:
-                self._data["ats"] = True
-
-    @property
-    def data(self):
-        """Return downgrade information."""
-        return self._data
-
-
-def perform_self_update():
-    """
-    Update files to latest release. Do nothing for Python package.
-
-    This function checks the latest GitHub release first.
-    If local version is not up-to-date, this function retrieves the latest
-    GitHub release asset (.tar.xz) and replaces existing files with extracted files.
-    """
-    # we don't update when Python package is used
-    try:
-        with open(os.path.join(os.path.dirname(Dir.scriptdir), "RELEASE")) as f:
-            current_release = f.readline().rstrip()
-    except Exception:
-        sys.exit("'RELEASE' file doesn't exist. Self update aborted.")
-
-    # get latest release
-    logging.info("Retrieving RELEASE from master")
-    try:
-        with urllib.request.urlopen(URL.release) as f:
-            release = f.readline().rstrip().decode("ascii")
-    except Exception as e:
-        sys.exit("Failed to retrieve RELEASE file: {}".format(e))
-
-    # do nothing if the installed version is latest
-    if release == current_release:
-        logging.info("Already up-to-date.")
-        return
-
-    # retrieve the release asset
-    archive_url = URL.rel_tarxz_tmpl.format(release)
-    logging.info("Retrieving release asset {}".format(archive_url))
-    try:
-        with urllib.request.urlopen(archive_url) as f:
-            asset_archive = f.read()
-    except Exception as e:
-        sys.exit("Failed to retrieve release asset file: {}".format(e))
-
-    # unpack the archive
-    logging.info("Unpacking archive {}".format(archive_url))
-    topdir = os.path.dirname(Dir.scriptdir)
-    try:
-        with tarfile.open(fileobj=io.BytesIO(asset_archive), mode="r:xz") as f:
-            f.extractall(topdir)
-    except Exception as e:
-        sys.exit("Failed to unpack release asset file: {}".format(e))
-
-    # update files
-    archive_dir = os.path.join(topdir, "truckersmp-cli-" + release)
-    for root, _dirs, files in os.walk(archive_dir, topdown=False):
-        inner_root = root[len(archive_dir):]
-        destdir = topdir + inner_root
-        logging.debug("Creating directory {}".format(destdir))
-        os.makedirs(destdir, exist_ok=True)
-        for f in files:
-            srcpath = os.path.join(root, f)
-            dstpath = os.path.join(destdir, f)
-            logging.info("Copying {} as {}".format(srcpath, dstpath))
-            os.replace(srcpath, dstpath)
-        os.rmdir(root)
-
-    # done
-    logging.info("Self update complete")
-
-
-def check_libsdl2():
-    """
-    Check whether SDL2 shared object file is present.
-
-    If SDL2 is detected, this function returns True.
-    Otherwise, this returns False.
-
-    Currently the check is done only on Linux systems.
-    """
-    if platform.system() != "Linux":
-        return True
-
-    try:
-        ctypes.cdll.LoadLibrary(File.sdl2_soname)
-        return True
-    except OSError:
-        return False
-
-
-def download_files(host, files_to_download, progress_count=None):
-    """Download files."""
-    file_count = progress_count[0] if progress_count else 1
-    num_of_files = progress_count[1] if progress_count else len(files_to_download)
-    conn = http.client.HTTPSConnection(host)
-    try:
-        while len(files_to_download) > 0:
-            path, dest, md5 = files_to_download[0]
-            md5hash = hashlib.md5()
-            bufsize = md5hash.block_size * 256
-            name = os.path.basename(path)
-            destdir = os.path.dirname(dest)
-            name_getting = "[{}/{}] Get: {}".format(file_count, num_of_files, name)
-            logging.debug(
-              "Downloading file https://{}{} to {}".format(host, path, destdir))
-
-            # make file hierarchy
-            os.makedirs(destdir, exist_ok=True)
-
-            # download file
-            conn.request("GET", path, headers={"Connection": "keep-alive"})
-            res = conn.getresponse()
-
-            if (res.status == 301
-                    or res.status == 302
-                    or res.status == 303
-                    or res.status == 307
-                    or res.status == 308):
-                # HTTP redirection
-                u = urllib.parse.urlparse(res.getheader("Location"))
-                if not download_files(
-                  u.netloc, [(u.path, dest, md5), ], (file_count, num_of_files)):
-                    return False
-                # downloaded successfully from redirected URL
-                del files_to_download[0]
-                file_count += 1
-                conn = http.client.HTTPSConnection(host)
-                continue
-            elif res.status != 200:
-                logging.error(
-                  "Server {} responded with status code {}.".format(host, res.status))
-                return False
-
-            lastmod = res.getheader("Last-Modified")
-            content_len = res.getheader("Content-Length")
-
-            with open(dest, "wb") as f:
-                downloaded = 0
-                while True:
-                    buf = res.read(bufsize)
-                    if not buf:
-                        break
-                    downloaded += len(buf)
-                    f.write(buf)
-                    md5hash.update(buf)
-                    if content_len:
-                        progress = "{:,} / {:,}".format(downloaded, int(content_len))
-                    else:
-                        progress = "{:,}".format(downloaded)
-                    print("\r{:40}{:>40}".format(name_getting, progress), end="")
-
-            if md5hash.hexdigest() != md5:
-                print("\r{:40}{:>40}".format(name, "MD5 MISMATCH"))
-                logging.error("MD5 mismatch for {}".format(dest))
-                return False
-
-            # wget-like timestamping for downloaded files
-            if lastmod:
-                timestamp = time.mktime(
-                  time.strptime(lastmod, "%a, %d %b %Y %H:%M:%S GMT")) - time.timezone
-                try:
-                    os.utime(dest, (timestamp, timestamp))
-                except Exception:
-                    pass
-
-            # downloaded successfully
-            print("\r{:40}{:>40}".format(name, "OK"))
-
-            # skip already downloaded files
-            # when trying to download from URL.dlurlalt
-            del files_to_download[0]
-
-            file_count += 1
-    except Exception as e:
-        logging.error("Failed to download https://{}{}: {}".format(host, path, e))
-        return False
-    finally:
-        conn.close()
-
-    return True
-
-
-def activate_native_d3dcompiler_47(prefix, wine):
-    """Download/activate native 64-bit version of d3dcompiler_47."""
-    # check whether DLL is already downloaded
-    md5hash = hashlib.md5()
-    need_download = True
-    try:
-        with open(File.d3dcompiler_47, "rb") as f:
-            while True:
-                buf = f.read(md5hash.block_size * 4096)
-                if not buf:
-                    break
-                md5hash.update(buf)
-        if md5hash.hexdigest() == File.d3dcompiler_47_md5:
-            logging.debug("d3dcompiler_47.dll is present, MD5 is OK.")
-            need_download = False
-    except Exception:
-        pass
-
-    # download 64-bit d3dcompiler_47.dll from ImagingSIMS' repo
-    # https://github.com/ImagingSIMS/ImagingSIMS
-    if need_download:
-        logging.debug("Downloading d3dcompiler_47.dll")
-        os.makedirs(Dir.dllsdir, exist_ok=True)
-        if not download_files(
-          URL.raw_github,
-          [(URL.d3dcompilerpath, File.d3dcompiler_47, File.d3dcompiler_47_md5), ]):
-            sys.exit("Failed to download d3dcompiler_47.dll")
-
-    # copy into system32
-    destdir = os.path.join(prefix, "drive_c/windows/system32")
-    logging.debug("Copying d3dcompiler_47.dll into {}".format(destdir))
-    shutil.copy(File.d3dcompiler_47, destdir)
-
-    # add DLL override setting
-    env = os.environ.copy()
-    env["WINEDEBUG"] = "-all"
-    env["WINEPREFIX"] = prefix
-    exename = "eurotrucks2.exe" if args.ets2 else "amtrucks.exe"
-    logging.debug("Adding DLL override setting for {}".format(exename))
-    subproc.call(
-      [wine, "reg", "add",
-       "HKCU\\Software\\Wine\\AppDefaults\\{}\\DllOverrides".format(exename),
-       "/v", "d3dcompiler_47", "/t", "REG_SZ", "/d", "native"],
-      env=env)
-
-
-def wait_for_steam(use_proton, loginvdf_paths, wine=None, env=None):
-    """
-    Wait for Steam to be running.
-
-    If use_proton is True, this function also detects
-    the Steam installation directory for Proton and returns it.
-
-    On user login the timestamp in
-    [steam installation directory]/config/loginusers.vdf gets updated.
-    We can detect the timestamp update with comparing timestamps.
-
-    use_proton: True if Proton is used, False if Wine is used
-    loginvdf_paths: loginusers.vdf paths
-    wine: Wine command (path or name)
-         (can be None if use_proton is True)
-    env: A dictionary that contains environment variables
-         (can be None if use_proton is True)
-    """
-    steamdir = None
-    loginusers_timestamps = []
-    for path in loginvdf_paths:
-        try:
-            st = os.stat(path)
-            loginusers_timestamps.append(st.st_mtime)
-        except OSError:
-            loginusers_timestamps.append(0)
-    if not check_steam_process(use_proton=use_proton, wine=wine, env=env):
-        logging.debug("Starting Steam...")
-        if use_proton:
-            subproc.Popen(
-              ("nohup", "steam"), stdout=subproc.DEVNULL, stderr=subproc.STDOUT)
-        else:
-            subproc.Popen(
-              ("nohup",
-               wine, os.path.join(args.wine_steam_dir, "steam.exe"), "-no-cef-sandbox"),
-              env=env, stdout=subproc.DEVNULL, stderr=subproc.STDOUT)
-        waittime = 99
-        while waittime > 0:
-            print(ngettext(
-              "\rWaiting {} second for steam to start up. ",
-              "\rWaiting {} seconds for steam to start up. ",
-              waittime).format(waittime), end="")
-            time.sleep(1)
-            waittime -= 1
-            for i, path in enumerate(loginvdf_paths):
-                try:
-                    st = os.stat(path)
-                    if st.st_mtime > loginusers_timestamps[i]:
-                        print("\r{}".format(" " * 70))  # clear "Waiting..." line
-                        logging.debug(
-                          "Steam should now be up and running and the user logged in.")
-                        steamdir = os.path.dirname(
-                          os.path.dirname(loginvdf_paths[i]))
-                        break
-                except OSError:
-                    pass
-            else:
-                continue
-            break
-        else:
-            # waited 99 seconds without detecting timestamp change
-            print("\r{}".format(" " * 70))
-            logging.debug("Steam should be up now.")
-            if use_proton:
-                # could not detect steam installation directory
-                # fallback to $XDG_DATA_HOME/Steam
-                steamdir = os.path.join(Dir.XDG_DATA_HOME, "Steam")
-    else:
-        # Steam is running
-        logging.debug("Steam is running")
-        if use_proton:
-            # detect most recently updated "loginusers.vdf" file
-            max_mtime = max(loginusers_timestamps)
-            for i, path in enumerate(loginvdf_paths):
-                if loginusers_timestamps[i] == max_mtime:
-                    steamdir = os.path.dirname(os.path.dirname(loginvdf_paths[i]))
-                    break
-    return steamdir
-
-
-def check_steam_process(use_proton, wine=None, env=None):
-    """
-    Check whether Steam client is already running.
-
-    If Steam is running, this function returns True.
-    Otherwise this returns False.
-
-    use_proton: True if Proton is used, False if Wine is used
-    wine: Wine command (path or name)
-         (can be None if use_proton is True)
-    env: A dictionary that contains environment variables
-         (can be None if use_proton is True)
-    """
-    if use_proton:
-        try:
-            subproc.check_call(
-              ("pgrep", "-u", getuser(), "-x", "steam"), stdout=subproc.DEVNULL)
-            return True
-        except Exception:
-            return False
-    else:
-        steam_is_running = False
-        argv = (wine, "winedbg", "--command", "info process")
-        env_wine = env.copy()
-        env_wine["WINEDLLOVERRIDES"] = "winex11.drv="
-        try:
-            output = subproc.check_output(argv, env=env_wine, stderr=subproc.DEVNULL)
-            for line in output.decode("utf-8").splitlines():
-                line = line[:-1]  # strip last "'" for rindex()
-                try:
-                    exename = line[line.rindex("'") + 1:]
-                except ValueError:
-                    continue
-                if exename.lower().endswith("steam.exe"):
-                    steam_is_running = True
-                    break
-        except subproc.CalledProcessError as e:
-            sys.exit("Failed to get Wine process list: " + e.output.decode("utf-8"))
-        return steam_is_running
-
 
 def start_with_proton():
     """Start game with Proton."""
-    steamdir = wait_for_steam(use_proton=True, loginvdf_paths=File.loginusers_paths)
+    steamdir = utils.wait_for_steam(use_proton=True,
+                                    loginvdf_paths=variables.File.loginusers_paths)
     logging.info("Steam installation directory: " + steamdir)
 
     if not os.path.isdir(args.prefixdir):
@@ -503,7 +38,8 @@ def start_with_proton():
     # activate native d3dcompiler_47
     wine = os.path.join(args.protondir, "dist/bin/wine")
     if args.activate_native_d3dcompiler_47:
-        activate_native_d3dcompiler_47(os.path.join(args.prefixdir, "pfx"), wine)
+        utils.activate_native_d3dcompiler_47(os.path.join(args.prefixdir, "pfx"),
+                                             wine, args.ets2)
 
     env = os.environ.copy()
     env["SteamGameId"] = args.steamid
@@ -516,7 +52,7 @@ def start_with_proton():
     if args.disable_proton_overlay:
         ld_preload = ""
     else:
-        overlayrenderer = os.path.join(steamdir, File.overlayrenderer_inner)
+        overlayrenderer = os.path.join(steamdir, variables.File.overlayrenderer_inner)
         if "LD_PRELOAD" in env:
             env["LD_PRELOAD"] += ":" + overlayrenderer
         else:
@@ -530,7 +66,7 @@ def start_with_proton():
         gamepath = os.path.join(args.gamedir, "bin/win_x64", exename)
         argv += gamepath, "-nointro", "-64bit"
     else:
-        argv += File.inject_exe, args.gamedir, args.moddir
+        argv += variables.File.inject_exe, args.gamedir, args.moddir
     logging.info("""Startup command:
   SteamGameId={}
   SteamAppId={}
@@ -558,17 +94,18 @@ def start_with_wine():
     """Start game with Wine."""
     wine = os.environ["WINE"] if "WINE" in os.environ else "wine"
     if args.activate_native_d3dcompiler_47:
-        activate_native_d3dcompiler_47(args.prefixdir, wine)
+        utils.activate_native_d3dcompiler_47(args.prefixdir, wine, args.ets2)
 
     env = os.environ.copy()
     env["WINEDEBUG"] = "-all"
     env["WINEARCH"] = "win64"
     env["WINEPREFIX"] = args.prefixdir
 
-    wait_for_steam(
+    utils.wait_for_steam(
       use_proton=False,
       loginvdf_paths=(os.path.join(args.wine_steam_dir, "config/loginusers.vdf"), ),
       wine=wine,
+      wine_steam_dir=args.wine_steam_dir,
       env=env,
     )
     if "WINEDLLOVERRIDES" not in env:
@@ -582,7 +119,7 @@ def start_with_wine():
         gamepath = os.path.join(args.gamedir, "bin/win_x64", exename)
         argv += gamepath, "-nointro", "-64bit"
     else:
-        argv += File.inject_exe, args.gamedir, args.moddir
+        argv += variables.File.inject_exe, args.gamedir, args.moddir
     logging.info("""Startup command:
   WINEDEBUG=-all
   WINEARCH=win64
@@ -595,380 +132,6 @@ def start_with_wine():
         logging.info("Wine output:\n" + output.decode("utf-8"))
     except subproc.CalledProcessError as e:
         logging.error("Wine output:\n" + e.output.decode("utf-8"))
-
-
-def update_mod():
-    """Download missing or outdated "multiplayer mod" files."""
-    if not os.path.isdir(args.moddir):
-        logging.debug("Creating directory {}".format(args.moddir))
-        os.makedirs(args.moddir, exist_ok=True)
-
-    # get the fileinfo from the server
-    try:
-        with urllib.request.urlopen(URL.listurl) as f:
-            files_json = f.read()
-    except Exception as e:
-        sys.exit("Failed to download files.json: {}".format(e))
-
-    # extract md5sums and filenames
-    modfiles = []
-    try:
-        for item in json.JSONDecoder().decode(str(files_json, "ascii"))["Files"]:
-            modfiles.append((item["Md5"], item["FilePath"]))
-        if len(modfiles) == 0:
-            raise Exception("File list is empty")
-    except Exception as e:
-        sys.exit("""Failed to parse files.json: {}
-Please report an issue: {}""".format(e, URL.issueurl))
-
-    # compare existing local files with md5sums
-    # and remember missing/wrong files
-    dlfiles = []
-    for md5, jsonfilepath in modfiles:
-        md5hash = hashlib.md5()
-        modfilepath = os.path.join(args.moddir, jsonfilepath[1:])
-        if not os.path.isfile(modfilepath):
-            dlfiles.append(("/files" + jsonfilepath, modfilepath, md5))
-        else:
-            try:
-                with open(modfilepath, "rb") as f:
-                    while True:
-                        buf = f.read(md5hash.block_size * 4096)
-                        if not buf:
-                            break
-                        md5hash.update(buf)
-                if md5hash.hexdigest() != md5:
-                    dlfiles.append(("/files" + jsonfilepath, modfilepath, md5))
-            except Exception as e:
-                sys.exit("Failed to read {}: {}".format(modfilepath, e))
-    if len(dlfiles) > 0:
-        message_dlfiles = "Files to download:\n"
-        for path, _, _ in dlfiles:
-            message_dlfiles += "  {}\n".format(path)
-        logging.info(message_dlfiles.rstrip())
-    else:
-        logging.debug("No files to download")
-
-    # download missing/wrong files
-    if not download_files(URL.dlurl, dlfiles):
-        if not download_files(URL.dlurlalt, dlfiles):
-            # something went wrong
-            sys.exit("Failed to download mod files.")
-
-
-def get_current_steam_user():
-    """
-    Get the current AccountName with saved login credentials.
-
-    If successful this returns the AccountName of the user with saved credentials.
-    Otherwise this returns None.
-
-    This function depends on the package "vdf".
-    """
-    loginvdf_paths = File.loginusers_paths.copy()
-    # try Wine Steam directory first when Wine is used
-    if args.wine:
-        loginvdf_paths.insert(0, os.path.join(args.wine_steam_dir, File.loginvdf_inner))
-    for path in loginvdf_paths:
-        try:
-            with open(path) as f:
-                login_vdf = vdf.parse(f)
-
-            for info in login_vdf["users"].values():
-                remember = "RememberPassword" in info and info["RememberPassword"] == "1"
-                recent_uc = "MostRecent" in info and info["MostRecent"] == "1"
-                recent_lc = "mostrecent" in info and info["mostrecent"] == "1"
-                if remember and (recent_lc or recent_uc) and "AccountName" in info:
-                    return info["AccountName"]
-        except Exception:
-            pass
-    return None
-
-
-def get_beta_branch_name(game_name="ets2"):
-    """
-    Get the current required beta branch name to comply with TruckersMP.
-
-    If downgrade is needed, this returns a branch name that can be used
-    to install TruckersMP-compatible versions of games (e.g. "temporary_1_36")
-    on success or None on error.
-    If downgrade is not needed, this returns None.
-    """
-    try:
-        parser = DowngradeHTMLParser()
-        with urllib.request.urlopen(URL.truckersmp_stats) as f:
-            parser.feed(f.read().decode("utf-8"))
-
-        if parser.data[game_name]:
-            version = get_supported_game_versions()[game_name].split(".")
-            return "temporary_{}_{}".format(version[0], version[1])
-        return None
-    except Exception:
-        return None
-
-
-def get_supported_game_versions():
-    """
-    Get TruckersMP-supported game versions via TruckersMP Web API.
-
-    Returns a dict of 'game: version' pairs:
-    {
-        "ets2": "1.36.2.55",
-        "ats": "1.36.1.40"
-    }
-    """
-    try:
-        with urllib.request.urlopen(URL.truckersmp_api) as f:
-            data = json.load(f)
-
-        return {
-            "ets2": data["supported_game_version"].replace("s", ""),
-            "ats": data["supported_ats_game_version"].replace("s", "")
-        }
-    except Exception:
-        return None
-
-
-def update_game():
-    """
-    Update game and Proton via SteamCMD.
-
-    We make sure Steam is closed before updating.
-    On Linux, we make sure both Windows and Linux version of Steam are closed.
-    It's possible to update with the Steam client open but the client looses
-    all connectivity and asks for password and Steam Guard code after restart.
-
-    When "--wine" is specified, this function retrieves/uses Windows version of
-    SteamCMD. When "--proton" is specified, this retrieves/uses
-    Linux version of SteamCMD.
-    """
-    steamcmd_prolog = ""
-    steamcmd_cmd = []
-
-    env = os.environ.copy()
-    env["WINEDEBUG"] = "-all"
-    env["WINEARCH"] = "win64"
-    env_steam = env.copy()
-    if args.proton:
-        # Proton's "prefix" is for STEAM_COMPAT_DATA_PATH that contains
-        # the directory "pfx" for WINEPREFIX
-        env_steam["WINEPREFIX"] = os.path.join(args.prefixdir, "pfx")
-    else:
-        env_steam["WINEPREFIX"] = args.prefixdir
-    # use a prefix only for SteamCMD to avoid every-time authentication
-    env["WINEPREFIX"] = Dir.steamcmdpfx
-    # don't show "The Wine configuration is being updated" dialog
-    # or install Gecko/Mono
-    env["WINEDLLOVERRIDES"] = "winex11.drv="
-
-    wine = env["WINE"] if "WINE" in env else "wine"
-    os.makedirs(Dir.steamcmdpfx, exist_ok=True)
-    try:
-        subproc.check_call((wine, "--version"), stdout=subproc.DEVNULL, env=env)
-        logging.debug("Wine ({}) is available".format(wine))
-    except subproc.CalledProcessError:
-        logging.debug("Wine is not available")
-        wine = None
-    if args.proton:
-        # we don't use system SteamCMD because something goes wrong in some cases
-        # see https://github.com/lhark/truckersmp-cli/issues/43
-        steamcmd = os.path.join(Dir.steamcmddir, "steamcmd.sh")
-        steamcmd_url = URL.steamcmdlnx
-        gamedir = args.gamedir
-    else:
-        if not wine:
-            sys.exit("Wine ({}) is not available.".format(wine))
-        steamcmd_prolog += """WINEDEBUG=-all
-  WINEARCH=win64
-  WINEPREFIX={}
-  WINEDLLOVERRIDES=winex11.drv=
-  {} """.format(Dir.steamcmdpfx, wine)
-
-        # steamcmd.exe uses Windows path, not UNIX path
-        try:
-            gamedir = subproc.check_output(
-              (wine, "winepath", "-w", args.gamedir), env=env).decode("utf-8").rstrip()
-        except Exception as e:
-            sys.exit(
-              "Failed to convert game directory to Windows path: {}".format(e))
-
-        steamcmd = os.path.join(Dir.steamcmddir, "steamcmd.exe")
-        steamcmd_cmd.append(wine)
-        steamcmd_url = URL.steamcmdwin
-    steamcmd_cmd.append(steamcmd)
-
-    # fetch SteamCMD if not in our data directory
-    os.makedirs(Dir.steamcmddir, exist_ok=True)
-    if not os.path.isfile(steamcmd):
-        logging.debug("Retrieving SteamCMD")
-        try:
-            with urllib.request.urlopen(steamcmd_url) as f:
-                steamcmd_archive = f.read()
-        except Exception as e:
-            sys.exit("Failed to retrieve SteamCMD: {}".format(e))
-        logging.debug("Extracting SteamCMD")
-        try:
-            if args.proton:
-                with tarfile.open(
-                  fileobj=io.BytesIO(steamcmd_archive), mode="r:gz") as f:
-                    f.extractall(Dir.steamcmddir)
-            else:
-                with ZipFile(io.BytesIO(steamcmd_archive)) as f:
-                    with f.open("steamcmd.exe") as f_exe:
-                        with open(steamcmd, "wb") as f_out:
-                            f_out.write(f_exe.read())
-        except Exception as e:
-            sys.exit("Failed to extract SteamCMD: {}".format(e))
-
-    logging.info("SteamCMD: " + steamcmd)
-
-    # Linux version of Steam
-    if platform.system() == "Linux" and check_steam_process(use_proton=True):
-        logging.debug("Closing Linux version of Steam")
-        subproc.call(("steam", "-shutdown"))
-    # Windows version of Steam
-    if wine and check_steam_process(use_proton=False, wine=wine, env=env_steam):
-        logging.debug("Closing Windows version of Steam in " + args.wine_steam_dir)
-        subproc.call(
-          (wine, os.path.join(args.wine_steam_dir, "steam.exe"), "-shutdown"),
-          env=env_steam)
-
-    if args.proton:
-        # download/update Proton
-        os.makedirs(args.protondir, exist_ok=True)
-        logging.debug("Updating Proton (AppID:{})".format(args.proton_appid))
-        logging.info("""Command:
-  {}
-    +login {}
-    +force_install_dir {}
-    +app_update {} validate
-    +quit""".format(steamcmd, args.account, args.protondir, args.proton_appid))
-        try:
-            subproc.check_call(
-              (steamcmd,
-               "+login", args.account,
-               "+force_install_dir", args.protondir,
-               "+app_update", str(args.proton_appid), "validate",
-               "+quit"))
-        except subproc.CalledProcessError:
-            sys.exit("SteamCMD exited abnormally")
-
-    # determine game branch
-    branch = "public"
-    if args.beta:
-        branch = args.beta
-    else:
-        game = "ats" if args.ats else "ets2"
-        beta_branch_name = get_beta_branch_name(game)
-        if beta_branch_name:
-            branch = beta_branch_name
-    logging.info("Game branch: " + branch)
-
-    # use SteamCMD to update the chosen game
-    os.makedirs(args.gamedir, exist_ok=True)
-    logging.debug("Updating Game (AppID:{})".format(args.steamid))
-    logging.info("""Command:
-  {}{}
-    +@sSteamCmdForcePlatformType windows
-    +login {}
-    +force_install_dir {}
-    +app_update {} -beta {} validate
-    +quit""".format(
-      steamcmd_prolog, steamcmd, args.account, gamedir, args.steamid, branch))
-    steamcmd_args = [
-        "+@sSteamCmdForcePlatformType", "windows",
-        "+login", args.account,
-        "+force_install_dir", gamedir,
-        "+app_update", args.steamid,
-        "-beta", branch,
-        "validate",
-        "+quit",
-    ]
-    try:
-        subproc.check_call(steamcmd_cmd + steamcmd_args, env=env)
-    except subproc.CalledProcessError:
-        sys.exit("SteamCMD exited abnormally")
-
-
-def check_args_errors():
-    """Check command-line arguments."""
-    # checks for updating and/or starting
-    if not args.update and not args.start:
-        logging.info("--update/--start not specified, doing both.")
-        args.start = True
-        args.update = True
-
-    # make sure only one game is chosen
-    if args.ats and args.ets2:
-        sys.exit("It's only possible to use one game at a time.")
-    elif not args.ats and not args.ets2:
-        logging.info("--ats/--ets2 not specified, choosing ETS2.")
-        args.ets2 = True
-
-    game = "ats" if args.ats else "ets2"
-    args.steamid = str(AppId.game[game])
-    if not args.prefixdir:
-        args.prefixdir = Dir.default_prefixdir[game]
-    if not args.gamedir:
-        args.gamedir = Dir.default_gamedir[game]
-
-    # checks for starting
-    if args.start:
-        # make sure proton and wine aren't chosen at the same time
-        if args.proton and args.wine:
-            sys.exit("Start with Proton (-p) or Wine (-w)?")
-        elif not args.proton and not args.wine:
-            if platform.system() == "Linux":
-                logging.info("Platform is Linux, use Proton")
-                args.proton = True
-            else:
-                logging.info("Platform is not Linux, use Wine")
-                args.wine = True
-
-    # make sure proton and wine are using the same default
-    if args.wine:
-        if (args.prefixdir == Dir.default_prefixdir["ats"]
-           or args.prefixdir == Dir.default_prefixdir["ets2"]):
-            logging.debug("""Prefix directory is the default while using Wine,
-making sure it's the same directory as Proton""")
-            args.prefixdir = os.path.join(args.prefixdir, "pfx")
-
-    # default Steam directory for Wine
-    if not args.wine_steam_dir:
-        args.wine_steam_dir = os.path.join(
-          args.prefixdir,
-          "" if args.wine else "pfx",
-          "dosdevices/c:/Program Files (x86)/Steam")
-
-    # checks for starting while not updating
-    if args.start and not args.update:
-        # check for game
-        if (not os.path.isfile(
-              os.path.join(args.gamedir, "bin/win_x64/eurotrucks2.exe"))
-            and not os.path.isfile(
-              os.path.join(args.gamedir, "bin/win_x64/amtrucks.exe"))):
-            sys.exit("""Game not found in {}
-Need to download (-u) the game?""".format(args.gamedir))
-
-        # check for proton
-        if not os.path.isfile(os.path.join(args.protondir, "proton")) and args.proton:
-            sys.exit("""Proton and no update wanted but Proton not found in {}
-Need to download (-u) Proton?""".format(args.protondir))
-
-    # checks for updating
-    if args.update and not args.account:
-        if vdf_is_available:
-            args.account = get_current_steam_user()
-        if not args.account:
-            logging.info("Unable to find logged in steam user automatically.")
-            sys.exit("Need the steam account name (-n name) to update.")
-
-    # info
-    logging.info("AppID/GameID: {} ({})".format(args.steamid, game))
-    logging.info("Game directory: " + args.gamedir)
-    logging.info("Prefix: " + args.prefixdir)
-    if args.proton:
-        logging.info("Proton directory: " + args.protondir)
 
 
 def create_arg_parser():
@@ -991,9 +154,9 @@ A working native Steam installation is needed for this.
 SteamCMD can use your saved credentials for convenience.
 """
     epilog = "Proton AppID list:\n"
-    for k, v in AppId.proton.items():
+    for k, v in variables.AppId.proton.items():
         default_mark = ""
-        if k == AppId.proton["default"]:
+        if k == variables.AppId.proton["default"]:
             default_mark += " (Default)"
         if k != "default":
             epilog += "    Proton {:13}: {:>10}{}\n".format(k, v, default_mark)
@@ -1022,9 +185,10 @@ SteamCMD can use your saved credentials for convenience.
               [Default: $XDG_DATA_HOME/truckersmp-cli/(Game name)/data]""")
     ap.add_argument(
       "-i", "--proton-appid", metavar="APPID", type=int,
-      default=AppId.proton[AppId.proton["default"]],
+      default=variables.AppId.proton[variables.AppId.proton["default"]],
       help="""choose a different AppID for Proton (Needs an update for changes)
-              [Default: {}]""".format(AppId.proton[AppId.proton["default"]]))
+              [Default:{}] """.format(variables.AppId.proton[
+                                        variables.AppId.proton["default"]]))
     ap.add_argument(
       "-l", "--logfile", metavar="LOG", type=str,
       default="",
@@ -1045,7 +209,7 @@ SteamCMD can use your saved credentials for convenience.
               with saved credentials)""")
     ap.add_argument(
       "-o", "--protondir", metavar="DIR", type=str,
-      default=Dir.default_protondir,
+      default=variables.Dir.default_protondir,
       help="""choose a different Proton directory
               [Default: $XDG_DATA_HOME/truckersmp-cli/Proton]
               While updating any previous version in this folder gets changed
@@ -1126,8 +290,8 @@ def main():
     # example:
     #     {"5.0": 1245040, "4.11": 1113280, "default": "5.0"}
     try:
-        with open(File.proton_json) as f:
-            AppId.proton = json.load(f)
+        with open(variables.File.proton_json) as f:
+            variables.AppId.proton = json.load(f)
     except Exception as e:
         sys.exit("Failed to load proton.json: {}".format(e))
 
@@ -1140,7 +304,8 @@ def main():
         version = ""
         try:
             # try to load "RELEASE" file for release assets or cloned git directory
-            with open(os.path.join(os.path.dirname(Dir.scriptdir), "RELEASE")) as f:
+            with open(os.path.join(os.path.dirname(variables.Dir.scriptdir),
+                                   "RELEASE")) as f:
                 version += f.readline().rstrip()
         except Exception:
             pass
@@ -1176,35 +341,35 @@ def main():
 
     # self update
     if args.self_update:
-        perform_self_update()
+        downloads.perform_self_update()
         sys.exit()
 
     # fallback to old local folder
     if not args.moddir:
-        if os.path.isdir(os.path.join(Dir.scriptdir, "truckersmp")):
+        if os.path.isdir(os.path.join(variables.Dir.scriptdir, "truckersmp")):
             logging.debug("No moddir set and fallback found")
-            args.moddir = os.path.join(Dir.scriptdir, "truckersmp")
+            args.moddir = os.path.join(variables.Dir.scriptdir, "truckersmp")
         else:
             logging.debug("No moddir set, setting to default")
-            args.moddir = Dir.default_moddir
+            args.moddir = variables.Dir.default_moddir
     logging.info("Mod directory: " + args.moddir)
 
     # check for errors
-    check_args_errors()
+    utils.check_args_errors(args)
 
     # download/update ATS/ETS2 and Proton
     if args.update:
         logging.debug("Updating game files")
-        update_game()
+        downloads.update_game(args)
 
     # update truckersmp when starting multiplayer
     if not args.singleplayer:
         logging.debug("Updating mod files")
-        update_mod()
+        downloads.update_mod(args.moddir)
 
     # start truckersmp with proton or wine
     if args.start:
-        if not check_libsdl2():
+        if not utils.check_libsdl2():
             sys.exit("SDL2 was not found on your system.")
         start_functions = (("Proton", start_with_proton), ("Wine", start_with_wine))
         i = 0 if args.proton else 1
